rename 'hook' to hooks, 'Context' to 'contexts' and update imports. Remove unused imports from Ticket.tsx
